+++ js/chatbot/shared/ChatBot.svelte
@@ -7,9 +7,11 @@
 		load_components,
 		get_components_from_messages
 	} from "./utils";
-	import type { NormalisedMessage } from "../types";
+	import type { NormalisedMessage, Option } from "../types";
 	import { copy } from "@gradio/utils";
+	import type { CopyData } from "@gradio/utils";
 	import Message from "./Message.svelte";
+	import { DownloadLink } from "@gradio/wasm/svelte";
 
 	import { dequal } from "dequal/lite";
 	import {
@@ -22,7 +24,13 @@
 	} from "svelte";
 	import { Image } from "@gradio/image/shared";
 
-	import { Clear, Trash, Community, ScrollDownArrow } from "@gradio/icons";
+	import {
+		Clear,
+		Trash,
+		Community,
+		ScrollDownArrow,
+		Download
+	} from "@gradio/icons";
 	import { IconButtonWrapper, IconButton } from "@gradio/atoms";
 	import type { SelectData, LikeData } from "@gradio/utils";
 	import type { ExampleMessage } from "../types";
@@ -40,6 +48,7 @@
 
 	export let _fetch: typeof fetch;
 	export let load_component: Gradio["load_component"];
+	export let allow_file_downloads: boolean;
 
 	let _components: Record<string, ComponentType<SvelteComponent>> = {};
 
@@ -106,6 +115,8 @@
 		share: any;
 		error: string;
 		example_select: SelectData;
+		option_select: SelectData;
+		copy: CopyData;
 	}>();
 
 	function is_at_bottom(): boolean {
@@ -186,7 +197,6 @@
 			dispatch("change");
 		}
 	}
-
 	$: groupedMessages = value && group_messages(value, msg_format);
 
 	function handle_example_select(i: number, example: ExampleMessage): void {
@@ -238,6 +248,14 @@
 			});
 		}
 	}
+
+	function get_last_bot_options(): Option[] | undefined {
+		if (!value || !groupedMessages || groupedMessages.length === 0)
+			return undefined;
+		const last_group = groupedMessages[groupedMessages.length - 1];
+		if (last_group[0].role !== "assistant") return undefined;
+		return last_group[last_group.length - 1].options;
+	}
 </script>
 
 {#if value !== null && value.length > 0}
@@ -258,9 +276,7 @@
 						dispatch("error", message);
 					}
 				}}
-			>
-				<Community />
-			</IconButton>
+			/>
 		{/if}
 		<IconButton Icon={Trash} on:click={() => dispatch("clear")} label={"Clear"}
 		></IconButton>
@@ -292,6 +308,14 @@
 								on:click={() => (is_image_preview_open = false)}
 								label={"Clear"}
 							/>
+							{#if allow_file_downloads}
+								<DownloadLink
+									href={image_preview_source}
+									download={image_preview_source_alt || "image"}
+								>
+									<IconButton Icon={Download} label={"Download"} />
+								</DownloadLink>
+							{/if}
 						</IconButtonWrapper>
 					</div>
 				{/if}
@@ -326,10 +350,30 @@
 					{show_copy_button}
 					handle_action={(selected) => handle_like(i, messages[0], selected)}
 					scroll={is_browser ? scroll : () => {}}
+					{allow_file_downloads}
+					on:copy={(e) => dispatch("copy", e.detail)}
 				/>
 			{/each}
 			{#if pending_message}
 				<Pending {layout} />
+			{:else}
+				{@const options = get_last_bot_options()}
+				{#if options}
+					<div class="options">
+						{#each options as option, index}
+							<button
+								class="option"
+								on:click={() =>
+									dispatch("option_select", {
+										index: index,
+										value: option.value
+									})}
+							>
+								{option.label || option.value}
+							</button>
+						{/each}
+					</div>
+				{/if}
 			{/if}
 		</div>
 	{:else}
@@ -430,7 +474,7 @@
 		align-items: center;
 		padding: var(--spacing-xl);
 		border: 0.05px solid var(--border-color-primary);
-		border-radius: var(--radius-xl);
+		border-radius: var(--radius-md);
 		background-color: var(--background-fill-secondary);
 		cursor: pointer;
 		transition: var(--button-transition);
@@ -611,5 +655,35 @@
 		display: flex;
 		justify-content: center;
 		align-items: center;
+	}
+
+	.options {
+		margin-left: auto;
+		padding: var(--spacing-xxl);
+		display: grid;
+		grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+		gap: var(--spacing-xxl);
+		max-width: calc(min(4 * 200px + 5 * var(--spacing-xxl), 100%));
+		justify-content: end;
+	}
+
+	.option {
+		display: flex;
+		flex-direction: column;
+		align-items: center;
+		padding: var(--spacing-xl);
+		border: 1px dashed var(--border-color-primary);
+		border-radius: var(--radius-md);
+		background-color: var(--background-fill-secondary);
+		cursor: pointer;
+		transition: var(--button-transition);
+		max-width: var(--size-56);
+		width: 100%;
+		justify-content: center;
+	}
+
+	.option:hover {
+		background-color: var(--color-accent-soft);
+		border-color: var(--border-color-accent);
 	}
 </style>