+++ js/annotatedimage/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @gradio/annotatedimage
 
+## 0.8.9
+
+### Dependency updates
+
+- @gradio/atoms@0.11.2
+- @gradio/utils@0.9.0
+- @gradio/statustracker@0.9.6
+- @gradio/upload@0.14.2
+
+## 0.8.8
+
+### Dependency updates
+
+- @gradio/atoms@0.11.1
+- @gradio/client@1.8.0
+- @gradio/utils@0.8.0
+- @gradio/upload@0.14.1
+- @gradio/statustracker@0.9.5
+
 ## 0.8.7
 
 ### Dependency updates