+++ js/tabs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @gradio/tabs
 
+## 0.3.5
+
+### Dependency updates
+
+- @gradio/utils@0.9.0
+
+## 0.3.4
+
+### Dependency updates
+
+- @gradio/utils@0.8.0
+
 ## 0.3.3
 
 ### Fixes